movie_test/requirements.txt
requests==2.22.0
Django==3.0
djangorestframework==3.10.3
django-filter==2.2.0
